+++ www/Repository/RoleRepository.php
@@ -1,0 +1,8 @@
+<?php
+namespace App\Repository;
+use App\Core\SQL;
+
+class RoleRepository extends AbstractRepository
+{
+
+}
