+++ www/application.yml
@@ -1,0 +1,1 @@
+environment : dev
